+++ engine/virt/virt.go
@@ -252,7 +252,7 @@ func (v *Virt) VirtualizationExecute(ctx context.Context, ID string, commands, e
 }
 
 // ResourceValidate validate resource usage
-func (v *Virt) ResourceValidate(ctx context.Context, cpu float64, cpumap map[string]int, memory, storage int64) error {
+func (v *Virt) ResourceValidate(ctx context.Context, cpu float64, cpumap map[string]int64, memory, storage int64) error {
 	// TODO list all containers, calcuate resource
 	return nil
 }
